requirements.txt
Flask==2.0.0
line-bot-sdk==1.19.0
pandas==1.2.4
dropbox==11.9.0








